requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
filelock==3.20.0
filetype==1.2.0
Flask==3.1.2
fonttools==4.60.1
fsspec==2025.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.3
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.27.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
openai==1.109.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pi_heif==0.22.0
pillow==10.4.0
pillow-avif-plugin==1.5.2
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.1.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
reportlab==4.2.5
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.8.1
roboflow==1.2.11
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
supervision==0.26.1
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.66.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.23
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
